fixed two coverity issues

+++ src/netem.c
@@ -136,7 +136,7 @@ int netem_get_params(char *iface, struct netem_params *params)
 		fprintf(stderr, "couldn't get loss for iface: %s\n", iface);
 		goto cleanup_qdisc;
 	}
-	params->loss = (double)loss / (UINT_MAX / 100);
+	params->loss = (double)loss / (double)(UINT_MAX / 100);
 
 	rtnl_qdisc_put(found_qdisc);
 	rtnl_qdisc_put(filter_qdisc);
